sdcard: refactor, move SPI-based SD card inner into a separate module
